DesarrolloCloudApiConverter/Subscriptor/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-pubsub==2.13.11
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
celery==5.0.5
redis==4.3.4